LibVideo/VP9: Fix data issues causing parse failures
The first pass at the BitStream's read_f implementation expected to
read only up to 8 bits of data, but since then it's been used to read
up to 32 bits of data, but it was never updated to support values that
large, causing data inaccuracies and failures. Two methods were also
added that were specialized for specific sizes commonly passed to
read_f, but they seem to be inaccurate, so they've been removed.

Additionally, this cleans up a couple other variables stored in wrong
types.